backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
et_xmlfile==2.0.0
mongoengine==0.29.1
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pymongo==4.12.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2